Sostituito yahoo finance con i dati di Baybit

+++ main.py
@@ -204,7 +204,7 @@ def fetch_history(symbol: str):
         ])
 
         # Converti i tipi e timestamp
-        df["timestamp"] = pd.to_datetime(df["timestamp"], unit="ms")
+        df["timestamp"] = pd.to_datetime(df["timestamp"].astype(int), unit="ms")
         df.set_index("timestamp", inplace=True)
         for col in ["Open", "High", "Low", "Close", "Volume"]:
             df[col] = pd.to_numeric(df[col], errors="coerce")
